requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
soundfile
torchmetrics
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
editdistance
pyctcdecode
kenlm
gdown
transformers
datasets
soundfile
wget
black
isort
pre-commit
